requirements.txt
python-chess==1.999
requests==2.32.5
